requirements.txt
click==8.1.7
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.2
platformdirs==4.2.1
python-dateutil==2.9.0.post0
pytubefix
yt-dlp
pytz==2024.1
scenedetect==0.6.3
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
imageio
pysrt==1.1.2
certifi==2024.2.2
charset-normalizer==3.3.2
ffmpy==0.3.2
idna==3.7
imageio-ffmpeg==0.4.9
imutils==0.5.4
joblib==1.4.2
katna==0.9.2
lazy_loader==0.4
networkx==3.3
opencv-contrib-python==4.10.0.82
packaging==24.0
pillow==10.3.0
psutil==5.9.8
requests==2.32.2
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
threadpoolctl==3.5.0
tifffile==2024.5.10
einops==0.8.0
filelock==3.14.0
fsspec==2024.5.0
huggingface-hub==0.23.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
PyYAML==6.0.1
regex==2024.5.15
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
transformers==4.41.1
triton==2.3.0
urllib3==2.2.1
pysrt==1.1.2
accelerate
bitsandbytes
nltk
flash-attn
sentencepiece
protobuf
fastapi[all]